requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.5
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
joblib==1.3.2
kombu==5.3.4
nltk==3.8.1
prompt-toolkit==3.0.41
python-dateutil==2.8.2
python-decouple==3.8
redis==5.0.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.10
pytz